+++ tests/test_mct_barenco_74_dirty.py
@@ -5,7 +5,6 @@ import pytest
 from qiskit.quantum_info.operators import Operator
 
 from quconot.implementations.mct_barenco_74_dirty import MCTBarenco74Dirty
-
 from tests.test_mct_base import BaseTestMCT
 
 
@@ -48,15 +47,6 @@ class TestMCTBarenco74Dirty(BaseTestMCT):
             return self._reverse_matrix_dict[controls_no]
         else:
             return self._matrix_dict[controls_no]
-
-    def _take_auxiliaries_no(self, controls_no: int):
-        if controls_no in self._auxiliary_dict:
-            return self._auxiliary_dict[controls_no]
-
-        mct = MCTBarenco74Dirty(controls_no)
-        self._auxiliary_dict[controls_no] = mct.num_auxiliary_qubits()
-
-        return self._auxiliary_dict[controls_no]
 
     def test_init(self):
         with pytest.raises(
