+++ tests/test_mct_no_auxiliary.py
@@ -5,7 +5,6 @@ import pytest
 from qiskit.quantum_info.operators import Operator
 
 from quconot.implementations.mct_no_auxiliary import MCTNoAuxiliary
-
 from tests.test_mct_base import BaseTestMCT
 
 
@@ -47,9 +46,6 @@ class TestMCTNoAuxiliary(BaseTestMCT):
             return self._reverse_matrix_dict[controls_no]
         else:
             return self._matrix_dict[controls_no]
-
-    def _take_auxiliaries_no(self, controls_no: int):
-        raise NotImplementedError
 
     def test_init(self):
         with pytest.raises(
